simplifying language on syntax errors
Tried to simplify the language in the header.  I also wanted to change this statement: "Won’t even try to run the program if it can’t be parsed" Python has to run the program up until that point to even see the error, but the script will shut down if the line of code can't be parsed.  So this statement is very confusing.

I also fussed with the wording to focus more on linguistic synonyms, which usually hits better with a librarian crowd.