Update variables.md--added faded f string challenge
I added an f string challenge to get students to practice f string syntax and it also illustrates how expressions can be used in an f string and builds on the variable and math in the above section. This is the first time I added a challenge to a lesson using GitHub, so I am not sure on the formatting.