Update libraries.md--change thing_name to function
I thought library_name.thing_name was a bit confusing so I suggest library_name.function. I also noticed function was misspelled a paragraph down. Another thing, the string function is not a library, but a module of the Python standard library, so technically it should be module.function. This is different from a library like pandas.