Update libraries.md--cleaning up confusion between libraries and modules
We say that modules and libraries are the same thing, but I don't think this is true. Modules are parts of libraries, and the reason the output about the string module says module is because we are looking at it in the Python Strandard Library. I am trying to clean this up.