Use append instead of re-assigning full list
In order to reinforce the introduction of append in the earlier lists episode, it might be smart to use it here and take advantage of the mutability of lists, rather than just re-creating the whole list.

The `print(odds)` line is optional and could be removed. It was added to reinforce visually the contents of the list after the append functions.